requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
playwright==1.47.0
readability-lxml==0.8.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
lxml[html_clean]==5.3.0
lxml_html_clean==0.4.1
httpx==0.27.2
